Update chapter-02.mm
The concept of the route in the decorator is explained for the first time in line73, (and further from line 77). Maybe, it would be easier to understand,  if you introduced it  around line 55, with the 'index' example: (something like "In web3py, the route for an action is built by the decorator. This differs from web3py, where the function name defines the route)
Line 247: is it one exception (auth) or 'some exceptions'? Which are the others?