Add DB_FOLDER description in _scaffold/settings.py
As a new user I disabled this because I was using MySql and didn't know it was needed, errors forced me to turn it back on, but I didn't know why.

Eventually found the description in the docs under "Other DAL constructor parameters".  If the change is not appropriate here, perhaps add it to the _examples app?