add labels dict to field form controls
Add a [labels] dict to the controls variable in FormStyleFactory.produce().  This is to provide functionality similar to web2py for custom forms and displaying the label defined when using Form with a field list (Similar to SQLFORM.factory() in web2py).  Now in a custom form in html I can:

[[=form.custom['begin'] ]]
<div>
[[=form.custom['labels']['field_name'] ]]
</div>
<div>
[[=form.custom['widgets']['field_name'] ]]
</div>

Prior to this change you could not use the 'labels' dict to get the label from the field when using custom forms.