Render hidden widgets for custom forms
Automatically render hidden_widgets for custom forms.  If I use [[=form]] in my template the hidden_widgets are rendered onto the page.  If I use [[=form.custom["being"] ]] then I need to render them manually.  This change renders the hidden_widgets automatically for custom forms.  I added the hidden widgets to the beginning of the form.  If preferred, I can change this to add them to the end of the form.