Render boolean form fields in line
When rendering boolean fields in a Form, render them in-line instead of label over control.  Do the same for the Delete checkbox and message.

I'm not sure if this is something you're interested in, but I feel it makes the form more readable.  Also, my implementation is a bit 'hackish'.  If you like the idea I'll see if I can come up with a better implementation.

I understand if you choose not to implement.