Ability to hide reference tables+checkbox display
This PR creates the ability to hide the reference dropdown menu on the create form and make it readonly on the edit form. 
You can also hide it on the edit form if you set the referenced db field value to readonly=False in addition.

To goal is if you program a form for a table that is referenced by another table and your user is not supposed to 
have the choice to decouple the initially set reference. 

So you dont want that dropdown created when you have a field reference.. in essence.. and replace it with a default value, 
but that should not be transparent to the user.

In addition this adds consistent display of boolean values when set to True as checkboxes instead of displaying the Text "True". 

The checkbox code came from Jim Steil.