Update login variables
Update login variables to default to empty strings in cases where the function is called but no value is sent in the request. py4web defaults the auth forms but I am sending these values via axios and they are not send in this case because they are still null since the user has not modified/edited the form but directly clicked "sign in".