Url_signer session independent
The line 124 in url_signer.py
https://github.com/web2py/py4web/blob/master/py4web/utils/url_signer.py#L124

print("on_request", self.session.get("_signature_key"))

It raises an error when the session paramenter is not inlcuded in URLsigner. 
So it is no posible to declare 
url_signer = URLSigner()  

or put a key parameter.
url_signer = URLSigner(key=URL_SIGNER_KEY)

Then it is possilbe to make url_signer session independent.