Added test runner action
Included the standard test runner boilerplate to allow the script to be executed as the main module. This ensures that the unit tests defined in the script are run when the script is executed directly.
To be caution, I'm adding it only to this test at the moment to limit any negative impact.